Shield types switched to double from int to fix game printing out 15 decimal places.

+++ WolcenEditor/CharacterDataObjects/InventoryEquipped.cs
@@ -24,9 +24,9 @@ namespace WolcenEditor
         public double DamageMin { get; set; }
         public double DamageMax { get; set; }
         public double ResourceGeneration { get; set; }
-        public int ShieldResistance { get; set; }
-        public int ShieldBlockChance { get; set; }
-        public int ShieldBlockEfficiency { get; set; }
+        public double ShieldResistance { get; set; }
+        public double ShieldBlockChance { get; set; }
+        public double ShieldBlockEfficiency { get; set; }
     }
 
     public class ItemArmor
